requirements.txt
# Core RAG dependencies
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0

# Google Gemini — called via REST API directly (avoids v1beta issues)
requests>=2.31.0

# Local embeddings — runs on your machine, no API key needed
sentence-transformers>=3.0.0

# PDF processing
pypdf>=4.3.0

# Vector store
numpy>=2.0.0
faiss-cpu>=1.13.0

# UI
streamlit>=1.38.0

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0

# Linting / formatting
ruff>=0.6.0
